extracted abstract drivetrain behavior to new abstract Drivetrain class
